django-hallo-app/requirements.txt
Django==3.1.6
psycopg2-binary==2.8.6
python-dotenv==0.17.0
Pillow==8.2.0
djangorestframework==3.12.4